fix: :arrow_down: fix(dependencies): Update dependencies and fix eslint version (#245)
Description:
This commit updates several dependencies in the project, including "@clerk/nextjs", "@tanstack/react-query", and "livekit-client". These updates are necessary to keep the project up-to-date with the latest features and security fixes. Additionally, the version of eslint in the devDependencies has been corrected from "9.4.0" to "^8.57.0" to ensure compatibility with other packages.

Details:
The following dependencies have been updated:
- "@clerk/nextjs"
- "@tanstack/react-query"
- "livekit-client"
The version of eslint in the devDependencies has been corrected from "9.4.0" to "^8.57.0".

Changeset:
- package.json
- package-lock.json

Modified files:
- package.json: The file where the dependencies of the project are declared. The updated packages have been reflected here.
- package-lock.json: This file is automatically generated for any operations where npm modifies either the node_modules tree, or package.json. The updates in the dependencies and the correction in the eslint version have been reflected here.

This commit ensures that the project is using the latest versions of its dependencies and that the eslint version is compatible with other packages.

Signed-off-by: RicardoGEsteves(Ricardo Esteves) 30535937+RicardoGEsteves@users.noreply.github.com

Refs: [#244]